+++ src/features/vote/usePriceRound.ts
@@ -13,7 +13,6 @@ export interface PriceRound {
 const signer = provider.getSigner();
 const contract = createVotingContractInstance(signer);
 
-console.log("contract", contract);
 // This can be accessed without logging the user in.
 export default function usePriceRound() {
   const [priceRounds, setPriceRounds] = useState<PriceRound[]>([]);
